refactor: move diff types into domain
