+++ extension/src/app/application/port/diff-cache.ts
@@ -1,4 +1,4 @@
-import type { DiffV2 } from "../../../types";
+import type { DiffV2 } from "../../domain/diff/types";
 
 export type DiffCachePort = {
   load(key: string): Promise<DiffV2 | undefined>;
