+++ extension/src/app/domain/diff/types.ts
