+++ extension/src/app/infrastructure/providers/fetch-queue.ts
@@ -1,4 +1,4 @@
-import { must } from "../../../util/must";
+import { must } from "../../presentation/util/must";
 import { RateLimitError } from "./github-client";
 
 type Job = {
