+++ extension/src/background/handler.test.ts
@@ -1,7 +1,7 @@
 import { describe, expect, it, vi } from "vitest";
 import { AuthError, type ChangedFile, RateLimitError } from "../app/infrastructure/providers/github-client";
-import type { DiffV2, GuidResolvedPush, SemanticDiffRequest } from "../types";
-import { must } from "../util/must";
+import type { DiffV2, GuidResolvedPush, SemanticDiffRequest } from "../app/domain/diff/types";
+import { must } from "../app/presentation/util/must";
 import type { DifferPort } from "../app/application/port/differ";
 import { DiffError } from "../app/infrastructure/providers/wasm-differ";
 import { createHandler, type Deps, type Handler } from "./handler";
