+++ extension/src/background/promiseCache.ts
@@ -1,4 +1,4 @@
-import { must } from "../util/must";
+import { must } from "../app/presentation/util/must";
 
 export type PromiseCache<V> = {
   get(key: string, compute: () => Promise<V>): Promise<V>;
