+++ extension/src/content/views.ts
@@ -1,4 +1,4 @@
-import type { DiffV2 } from "../types";
+import type { DiffV2 } from "../app/domain/diff/types";
 
 // json is mutated in place by the push listener (merge resolved / replace on final push)
 export type ViewEntry = {
